back-end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["LaurieHC <lhumphriescable@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
Django = "3.2.11"
djangorestframework = "3.14"
django-filter = "23.2"
django-cors-headers = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
